python/query: Only #include <pygobject> if building with SWDB
Mixing hand-generated Python bindings with pygobject gets messy;
it looks like that's what this code is trying to do.  We don't
need this though unless swdb is enabled.

Today rpm-ostree doesn't use the libdnf python bindings at all; we build with
them on because it's not worth writing a patch to turn them off. But having this
`#include` introduces a new build dependency which I'd like to avoid.