tests: Link executable with -fPIC
I'm trying to reproduce a build failure for rpm-ostree
with Fedora's `LDFLAGS` which look like e.g.:
`export LDFLAGS='-Wl,-z,relro`
But I ended up hitting this issue (not the one I was looking for).
I didn't fully verify this but I believe the problem here is
we're building the `testshared.cpp` twice, and it needs to be
consistently `PIC`.